feat: Add prime factorization in Noir
Compared to C++, only puny numbers can be factorized in Noir in a
reasonable time (order of seconds). The program in its current state
(factorizing numbers less than 65536) takes ~12s for `nargo prove` and
~4s for `nargo execute` on a 32-core gen 13 i9 with 32G of RAM.

For comparison, the C++ program can factorize numbers up to
20'000'000'000'000'003 in comparable time. This one is a prime, which
leads to worst-case runtime, and it takes ~18s. 20'000'000'000'000'007
is factorized in 0.3s, since its composite.

Not that any of that matters, though.