refactor(docs.ws): Replace approach for generating mdx files
We previously used `ReactDOMServer.renderToStaticMarkup` to render
React components as static strings, which was prone to hydration issues
and setup complexity. To address these problems, we've switched to a new
approach where we call the React component for each source unit directly
in the MDX file, passing relevant data as a JSON string. The generator
produces MDX file with needed import and component calls.

Co-authored-by: Emil Ivanichkov <emil.ivanichkov@gmail.com>