fix(sol-reflector): Fix missing `functionSelector` field
This change is necessary because `FunctionDocItem` is a class,
and the constructor needs to know how to handle the `functionSelector`
field when it's called. For more details on why this is essential,
please refer to the `extractFields` function.

The change of `_modifiers` follows the same logic but in the opposite
direction. It's included in this commit because it's too small a change
to live on its own