config(base-utils): Add `rollup` configuration
This commit introduces Rollup as the primary building tool for the lib. Rollup
is a flexible and efficient module bundler. It provides excellent support for
both CommonJS and ECMAScript Modules (ESM). The decision to switch to Rollup was
made to enhance the project's build process, improve performance, and enable
seamless usage of modern JavaScript features.

Benefits of Using Rollup:
- The support for ESM and CJS formats.
- Rollup provides faster build times compared to other bundlers. This makes the
  development process more efficient.
- With the proper configuration, Rollup handles dependencies and external
  packages smoothly. This reduces potential issues with missing/broken imports.

Details of Changes:
- Added the Rollup configuration file (`rollup.config.js`). This configuration
  file specifies the build settings, such as format (ESM/CJS), input file,
  output directories etc.
- Updated the TypeScript configuration (`tsconfig.json`) to match Rollup's
  output settings.
- Update the `exports` field of `package.json` to point to the correct build
  artifacts.