feat(data-feeds-config-generator): Sort data feeds by description
Note that we first use keccak to hash the description, so that the order
is not obvious. The side effect is that feed id 0 is FOXY. Which is
good.

This will make the list a little more stable between PRs, until we
implement a smarter system to track addition and removal of feeds.