feat(sequencer): Report only BTC/USD and ETH/USD on holesky
This is a temporary solution until we have a way to blacklist feeds on
each network.

This is needed for the ongoing hackathon at Citrea. Setting for Holesky
to test out the setup and to wait for 24hrs to pass from the initial
deployment on Citrea. Then, we'll switch the "special_network" name to
Citrea.

Co-authored-by: Emil Ivanichkov <emil.ivanichkov@gmail.com>