refactor(DFCG): Refactor getAssetMetadata
`getAssetMetadata` now handles feeds with multiple resources per
exchange.

Use case: For example, consider a data feed `X / USD`. If an exchange
supports any of the pairs `X / USD`, `X / USDT`, or `X / USDC`, we want
to include the information for all of them in the aggregated provider
info.

The function filters the supported assets based on the exchange's
supported pairs and then reduces the data into a single object
(`providerInfo`). Each key in this object corresponds to a key found in
the `data` property of the assets, with values being arrays that
aggregate all the values associated with that key across all assets. If
no data is aggregated, the function returns `null`; otherwise, it
returns the `providerInfo` object.

Use case: