fix(DFCG): Fix unexpected behaviour in `getAllProvidersForFeed`
There were weird problem when the provider data were wrong.
We noticed incorrect values for the stable variants data and
overwrites.

In a nutshell the solution is to always start from empty `providers`
in this function. After all it takes care to create the providers
object for a certain feed from scratch.