feat(nix/modules): Replace `oracle-script-wasm` with `package`
The option was made redundant by the `id` option, furthermore it didn't
allow overriding the implementation.

Set the `package`'s default value to the wasm file in the derivation
corresponding to the oracle script.

Stop copying wasm files to working directory in `process-compose` and
`systemd` as now the blocksense config contains the absolure path to the
wasm files.