fix(reporter): Make `BlocksenseConfig` fields `Option` rather than `serde(default)`
Default values are obviously error prone but another problem this solves
is that it was impossible to not set `kafka_endpoint` in the reporter.
This is because all fields had `serde(default)` which assigns
`kafka_endpoint` to an empty string if it's missing. This was propagated
as an empty string in the `spin.toml`. This is wrong since the reporter
performs some `Option` checks on the `kafka_endpoint`.