build(direnv): Exit when nix-direnv fallback is triggered
Adds a check to `.envrc` to explicitly exit direnv if the
`NIX_DIRENV_DID_FALLBACK` variable is set. This ensures that the shell
environment is not loaded in an inconsistent or incorrect state when
`nix-direnv` cannot properly set up the Nix environment.

This change addresses issues where the developer environment might
appear to load but is missing crucial Nix-provided tools or
configurations due to a silent fallback. For example when the
devshell fail to build.

Ref: https://github.com/nix-community/nix-direnv/pull/487