feat(sequencer): Add round counters per provider
* Reading round counters from adfs contract and latest published values
* Increase round counters in provider after reading them from chain
* Add test that reads round values from chain
* Refactor how smart contracts are configured in sequencer config
* Fix(sequencer/config): Configuration related to contracts for second round consensus made consistent
* Fix example for process compose
* Use function pointers and chunked calls to multicall contract
* Fix changes that were causing a mismatch between sequencer and reporter when generating calldata
* Fix(e2e-tests): Fix failing tests due to changes in the config and process compose setup
* Add config option for reading values and round counteres
* feat(sequencer/providers): Read data from networks in parallel
* Panic when we can't load history from chain

To speed up initialization in the sequencer we read data in parallel from
all networks. We also change the readiness_probe to not have a 120
second initial delay, but probe every 10 seconds and require 10
consecutive failing probes in order to restart the process.

Co-authored-by: Hristo Staykov <izo8611@gmail.com>
Co-authored-by: Aneta Tsvetkova <anetastsvetkova@gmail.com>

feat(example-setup-03): Set `should-load-round-counters = false`

This will be set to keep the current state of the e2e tests as they
currently are. Once we update them to work with the rounds parsing
logic of the sequencer we will remove it