fix(config-types): Use `discriminated union` for oracle arguments schema
In our `additional_feed_info` schema, the `arguments` field defines the
settings passed to each oracle. Previously, this field was defined as a
`Schema.Union`. According to Effect's documentation \[1], **the order of
schemas in a union matters**—if multiple schemas can decode the same
value, the first matching one is chosen.

This became problematic because `cexPriceFeedsArgsSchema` allows both
fields to be optional, making it too permissive and causing incorrect
matches with other oracles' argument schemas. This led to subtle bugs
and misinterpretations during runtime.

To resolve this, we now use a **discriminated union** \[2] with a `kind`
field that explicitly specifies which oracle the arguments are intended
for.

* Update oracle schemas in
  `libs/ts/config-types/src/data-feeds-config/oracles.ts`
* Updated `feeds_config_v2.json`
* Updated `spin.toml` files for all oracles
* Fixed argument parsing logic in the `gecko-terminal` oracle

* \[1] [https://effect.website/docs/schema/basic-usage/#union-member-evaluation-order](https://effect.website/docs/schema/basic-usage/#union-member-evaluation-order)
* \[2] [https://effect.website/docs/schema/basic-usage/#discriminated-unions](https://effect.website/docs/schema/basic-usage/#discriminated-unions)