refactor(docs.ws): Add `networkNameToIconName` util function
At this point all network icons match the first part of the network name
(e.g: for ethereum-mainnet, ethereum-sepolia, etc. the icon is the same
and is called `ethereum`). But there are networks where for mainnet and
testnet the used icons are different. Extracting the logic for `network
name` <-> `network icon` match to a separate function will help for
future iterations where the current pattern does not work.